Fix treatment of good days, if no request was logged
Good days that got no request logged had their data point removed, which lead
to the day not been drawn for that version. See February 15 on

  http://gp.wmflabs.org/graphs/stc_al_jawal_saudi_arabia_daily_wikipedia_page_requests_by_version

previously to this commit. There the “M” version is drawn for February 15,
while the “Z” version is marked as NaN. We now properly mark absence of
requests for good days as seeing 0 requests.