Strip markup from MCC-MNC loaded from wiki table
The mapping between the carrier configuration and the zero configuration from
the wiki table at [1] relies on the MCC-MNCs. This mapping got broken, as the
MCC-MNC column of [1] got decorated with markup linking to the corresponding
Zero configuration on meta since 2013-11-10. We now strip this markup again
on our side, to unbreak this mapping.

[1] https://wikimediafoundation.org/w/index.php?title=Mobile_partnerships&oldid=94275#Where_is_Wikipedia_free_to_access.3F